+++ handlers/cookies.go
@@ -9,16 +9,15 @@ import (
 	guuid "github.com/google/uuid"
 )
 
-// Création du cookie. Renvoie la valeur du cookie.
-// Fonction appelée lorsqu'un utilisateur se connecte.
+// Permet de créer un cookie. Fonction appelée lorsqu'un utilisateur se connecte.
 func CreateCookie(w http.ResponseWriter, r *http.Request, username string) error {
 	id := guuid.New()
 	start := time.Now()
-	start2 := start.Add(time.Minute * 15)
+	end := start.Add(time.Minute * 15)
 	c := http.Cookie{
 		Name:    "CookieSession",
 		Value:   id.String(),
-		Expires: start2}
+		Expires: end}
 	http.SetCookie(w, &c)
 
 	// Ajout de l'UUID dans la base de données.
@@ -26,7 +25,6 @@ func CreateCookie(w http.ResponseWriter, r *http.Request, username string) error
 }
 
 // Renvoie la valeur du cookie actif. Renvoie une erreur lorsqu'il n'y a plus de cookie.
-// Utilisé lors du chargement d'une page
 func ReadCookie(w http.ResponseWriter, r *http.Request, name string) string {
 	c, err := r.Cookie(name)
 	if err != nil {
@@ -38,20 +36,20 @@ func ReadCookie(w http.ResponseWriter, r *http.Request, name string) string {
 	return c.Value
 }
 
-// Retourne le nom de l'utilisateur connecté. Utilise la fonction ReadCookie pour savoir sa valeur.
-// Parcourt la BDD pour savoir quel utilisateur est lié au cookie de session.
-// Utilisé à chaque chargement de page.
+// Permet de savoir si un utilisateur est connecté et son pseudo
 func VerifyUserConnected(w http.ResponseWriter, r *http.Request) UserConnectedStruct {
 	var userConnected UserConnectedStruct
+	// On lit le cookie de la page pour avoir sa valeur
 	CookieValue := ReadCookie(w, r, "CookieSession")
 	if CookieValue == "" {
 		userConnected.PseudoConnected = ""
 		userConnected.Connected = false
 		return userConnected
 	}
-
 	var user_connected string
 	db := BDD.OpenDataBase()
+
+	// On parcourt la BDD pour savoir quel utilisateur est lié au cookie de session.
 	checkUUID, err := db.Prepare("SELECT user_pseudo FROM session WHERE UUID = ?")
 	if err != nil {
 		log.Fatal(err)
@@ -65,13 +63,14 @@ func VerifyUserConnected(w http.ResponseWriter, r *http.Request) UserConnectedSt
 		check_user.Scan(&user_connected)
 	}
 	checkUUID.Close()
+
+	// On retourne le pseudo et le status de la connexion
 	userConnected.PseudoConnected = user_connected
 	userConnected.Connected = true
 	return userConnected
 }
 
-// Supprime le cookie actif
-// Fonction appelée lorsque l'utilisateur souhaite se déconnecter
+// Permet de supprimer un cookie lors de la déconnexion voulue d'un utilisateur
 func GetDeconnected(w http.ResponseWriter, r *http.Request) {
 	c := http.Cookie{
 		Name:   "CookieSession",
